third_party/blink/web_tests/fast/parser/foster-parent-expected.txt

Tests the HTML foster parenting algorithm when a DocumentFragment is the chosen foster parent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS div.children.length is 0
PASS docFragment.firstElementChild.tagName is 'H1'
PASS successfullyParsed is true

TEST COMPLETE
